spanner-lib/pyproject.toml
[tool.poetry]
name = "quizlet"
version = "0.1.0"
description = "A digital flashcards system using Google Cloud Spanner"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
google-cloud-spanner = "^3.40.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"